Added needed packages for Retrofit
