+++ SafeGarage_Mobile/app/src/main/java/com/quickreports/safegarage_mobile/activity/fragments/AlarmFragment.java
