+++ SafeGarage_Mobile/app/src/main/java/com/quickreports/safegarage_mobile/models/Message.java
@@ -1,0 +1,4 @@
+package com.quickreports.safegarage_mobile.models;
+
+public class Message {
+}
