+++ SafeGarage_Mobile/app/src/main/res/layout/fragment_alarm.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="fill_parent"
     android:layout_height="fill_parent"
-    tools:context=".fragments.AlarmFragment">
+    tools:context=".activity.fragments.AlarmFragment">
 
     <GridLayout
         android:layout_width="fill_parent"
