+++ SafeGarage_Mobile/app/src/main/res/layout/fragment_time.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.TimeFragment">
+    tools:context=".activity.fragments.TimeFragment">
 
     <LinearLayout
         android:layout_width="wrap_content"
