+++ SafeGarage_Mobile/app/src/main/res/layout/main_activity.xml
@@ -21,7 +21,7 @@
 
         <fragment
             android:id="@+id/pair_fragment"
-            class="com.quickreports.safegarage_mobile.fragments.PairFragment"
+            class="com.quickreports.safegarage_mobile.activity.fragments.PairFragment"
             android:layout_width="fill_parent"
             android:layout_height="fill_parent" />
 
@@ -29,7 +29,7 @@
 
     <fragment
         android:id="@+id/door_fragment"
-        class="com.quickreports.safegarage_mobile.fragments.DoorFragment"
+        class="com.quickreports.safegarage_mobile.activity.fragments.DoorFragment"
         android:layout_width="fill_parent"
         android:layout_height="match_parent"
         android:layout_weight="50"/>
